+++ broker/src/main/java/io/zeebe/broker/clustering/topology/TopologyManagerImpl.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.0. You may not use this file
  * except in compliance with the Zeebe Community License 1.0.
  */
-package io.zeebe.broker.clustering.base.topology;
+package io.zeebe.broker.clustering.topology;
 
 import static io.zeebe.broker.Broker.actorNamePattern;
